requirements.txt
fastapi
uvicorn
duckdb
asyncio
cachetools
pymupdf
pandas
httpx
openai
pythainlp
python-dotenv
qdrant-client
sentence-transformers
